[XNP-6120] Preset setup > Motion detection / Type text

+++ app/scripts/controllers/setup/analytics/motionDetectionController.js
@@ -2357,29 +2357,39 @@ kindFramework.controller('motionDetectionCtrl', function ($scope, $rootScope, Su
         });
     }
 
+    function changePresetType(newVal, oldVal){
+        if(newVal === 'Global'){
+            gotoPreset('Stop', $scope.presetData.oldPreset);
+        }else{
+            gotoPreset('Start', $scope.presetData.preset);
+        }
+    }
+
     $scope.$watch('presetData.type',function(newVal, oldVal){
         if(oldVal === null){
             $scope.presetData.oldType = newVal;
         }else{
-            // console.info(456);
-            // console.info(newVal, oldVal);
             $scope.presetData.oldType = oldVal;
 
-            if(validatePage()){
-                $rootScope.$emit('changeLoadingBar', true);
-                sketchbookService.removeDrawingGeometry();
-                saveSettings().finally(
-                    function(){
-                        if(newVal === 'Global'){
-                            gotoPreset('Stop', $scope.presetData.oldPreset);
-                        }else{
-                            gotoPreset('Start', $scope.presetData.preset);
+            if($scope.MotionDetection.MotionDetectionEnable === true){
+                if(validatePage()){
+                    $rootScope.$emit('changeLoadingBar', true);
+                    sketchbookService.removeDrawingGeometry();
+                    saveSettings().finally(
+                        function(){
+                            changePresetType(newVal, oldVal);
                         }
-                    }
-                );
+                    );
+                }
+            }else{
+                changePresetType(newVal, oldVal);
             }
         }
     });
+
+    function changePreset(newVal, oldVal){
+        gotoPreset('Start', newVal, true, oldVal);
+    }
 
     $scope.$watch('presetData.preset',function(newVal, oldVal){
         if(oldVal === null){
@@ -2389,14 +2399,18 @@ kindFramework.controller('motionDetectionCtrl', function ($scope, $rootScope, Su
             $scope.presetData.oldType = 'Preset';
             $scope.presetData.oldPreset = oldVal;
 
-            if(validatePage()){
-                $rootScope.$emit('changeLoadingBar', true);
-                sketchbookService.removeDrawingGeometry();
-                saveSettings().finally(
-                    function(){
-                        gotoPreset('Start', newVal, true, oldVal);
-                    }
-                );
+            if($scope.MotionDetection.MotionDetectionEnable === true){
+                if(validatePage()){
+                    $rootScope.$emit('changeLoadingBar', true);
+                    sketchbookService.removeDrawingGeometry();
+                    saveSettings().finally(
+                        function(){
+                            changePreset(newVal, oldVal);
+                        }
+                    );
+                }
+            }else{
+                changePreset(newVal, oldVal);
             }
         }
     });
